% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aic}
\alias{aic}
\title{AIC}
\usage{
aic(model_fitted, estimate_likelihood = TRUE, min_ess = 100, loglik = NULL)
}
\arguments{
\item{model_fitted}{A fitted multistate models object}

\item{estimate_likelihood}{logical; whether to estimate the log-likelihood, defaults to TRUE.}

\item{min_ess}{minimum effective sample size per subject, defaults to 100.}

\item{loglik}{value of the log-likelihood to use, if provided.}
}
\value{
An integer with calculated AIC
}
\description{
Akaike's Information Criterion, defined as -2\emph{log(L) + 2}k,
where L is the likelihood and k is the number of consumed degrees of freedom.
}
