% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cumulative_incidence}
\alias{cumulative_incidence}
\title{Cumulative incidence}
\usage{
cumulative_incidence(t, model, multistatemodels, subj = 1)
}
\arguments{
\item{t}{time or a vector of times}

\item{model}{multistate model}

\item{multistatemodels}{Loaded multistatemodels Julia environment}

\item{subj}{subject id}
}
\value{
A matrix with the cumulative incidence where columns represent each
possible transition and rows represent t
}
\description{
Compute the cumulative incidence for each possible transition as a function of time since state entry.
Assumes the subject starts their observation period at risk and saves cumulative incidence at the supplied vector of times, t.
}
\examples{
cumulative_incidence(t=2, model=model, multistatemodels=multistatemodels)
}
