% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_loglik}
\alias{estimate_loglik}
\title{Estimate log-likelihood}
\usage{
<<<<<<< HEAD
estimate_loglik(model, min_ess = 100)
=======
estimate_loglik(model, multistatemodels, min_ess = 100)
>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
}
\arguments{
\item{model}{multistate model}

<<<<<<< HEAD
=======
\item{multistatemodels}{Loaded multistatemodels Julia environment}

>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
\item{min_ess}{minimum effective sample size, defaults to 100.}
}
\value{
A Julia environment with a named Tuple
}
\description{
Estimate the log marginal likelihood for a fitted multistate model.
Require that the minimum effective sample size per subject is greater than min_ess.
}
\examples{
<<<<<<< HEAD
estimate_loglik(model=model)
=======
estimate_loglik(model=model, multistatemodels=multistatemodels)
>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
}
