% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_surrogate}
\alias{fit_surrogate}
\title{Fit a Markov surrogate model}
\usage{
fit_surrogate(
  model,
  surrogate_parameters = NULL,
  surrogate_constraints = NULL,
  crude_inits = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{model}{A model created in the multistatemodel function}

\item{surrogate_parameters}{Surrogate parameters}

\item{surrogate_constraints}{Surrogate constraints on model parameters}

\item{crude_inits}{logical; Defaults to TRUE}

\item{verbose}{logical; print messages, defaults to TRUE}
}
\value{
A Julia environment with a fitted Markov surrogate model
}
\description{
Fit a Markov surrogate model
}
\examples{
model_fitted_surrogate <- fit_surrogate(model=model)
}
