% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_convergence_records}
\alias{get_convergence_records}
\title{Get Convergence Records}
\usage{
get_convergence_records(model_fitted)
}
\arguments{
\item{model_fitted}{A fitted multistate models object}
}
\value{
A Julia environment with status, candidate solution, algorithm,
convergence records, and work counters
}
\description{
Return the convergence records for the fit.
}
