% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_loglik}
\alias{get_loglik}
\title{Get Log-Likelihood}
\usage{
get_loglik(model_fitted, multistatemodels, ll = "loglik")
}
\arguments{
\item{model_fitted}{A fitted multistate models object}

\item{multistatemodels}{Loaded multistatemodels Julia environment}

\item{ll}{string; one of "loglik" (default) for the observed data log-likelihood or
"subj_lml" for log marginal likelihood at the subject level}
}
\value{
An integer with the log-likelihood at the MLE
}
\description{
Return the log-likelihood at the maximum likelihood estimates.
}
\examples{
get_loglik(model_fitted=model_fitted, multistatemodels=multistatemodels)
}
