% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_parnames}
\alias{get_parnames}
\title{Get Parameter Names}
\usage{
get_parnames(model_fitted)
}
\arguments{
\item{model_fitted}{A fitted multistate models object}
}
\value{
A matrix with parameter names as strings
}
\description{
Return the parameter names
}
