% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initialize_parameters}
\alias{initialize_parameters}
\title{Initialize parameters}
\usage{
initialize_parameters(
  model,
  constraints = NULL,
  surrogate_constraints = NULL,
  surrogate_parameters = NULL,
  crude = FALSE
)
}
\arguments{
\item{model}{A model created in the multistatemodel function}

\item{constraints}{Constraints on model parameters}

\item{surrogate_constraints}{Surrogate constraints}

\item{surrogate_parameters}{Surrogate parameters}

\item{crude}{logical; Defaults to FALSE}
}
\description{
Modify the parameter values in a MultistateProcess object,
calibrate to the MLE of a Markov surrogate.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
model_fit <- initialize_parameters(model = model_fit)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
