% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistatemodel}
\alias{multistatemodel}
\title{Construct a multistate model}
\usage{
multistatemodel(
  hazard,
  data,
  SamplingWeights = NULL,
  CensoringPatterns = NULL,
  verbose = FALSE
)
}
\arguments{
\item{hazard}{A vector of cause specific hazards}

\item{data}{Data frame}

\item{SamplingWeights}{Sampling weights}

\item{CensoringPatterns}{Censoring patterns}

\item{verbose}{logical; print messages, defaults to FALSE}
}
\value{
A Julia MultistateModel environment that can be used for simulation and inference
}
\description{
Construct a multistate model
}
