% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistatemodels_load}
\alias{multistatemodels_load}
\title{Load Julia}
\usage{
multistatemodels_load(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
Load needed Julia packages. Run to use multistatemodels.
}
\examples{
multistatemodels <- multistatemodels_load()
}
