% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistatemodels_setup}
\alias{multistatemodels_setup}
\title{Setup Multistate Models with Julia packages}
\usage{
multistatemodels_setup()
}
\description{
This function should be run once after first installing the R package, and
any time there are updates to the Julia MultistateModels package.
This allows the necessary Julia packages to be installed before they can be used.
}
\examples{
multistatemodels_setup()
}
