% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_parameters}
\alias{set_parameters}
\title{Set parameters}
\usage{
set_parameters(model, newvalues)
}
\arguments{
\item{model}{multistate model}

\item{newvalues}{A list, each element of the list is new parameters for each hazard}
}
\description{
Set model parameters given a vector of values.
Copies newvalues to model.parameters.
}
