% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate}
\alias{simulate}
\title{Simulate data}
\usage{
simulate(
  model,
  nsim = 1,
  data = TRUE,
  paths = FALSE,
  delta_u = sqrt(.Machine$double.eps),
  delta_t = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{model}{multistate model}

\item{nsim}{number of sample paths to simulate}

\item{data}{logical; if TRUE then return discretely observed sample paths}

\item{paths}{logical; if FALSE then continuous-time sample paths not returned}

\item{delta_u}{minimum cumulative incidence increment per-jump, defaults to the larger of (1 / N subjects)^2 or sqrt(eps())}

\item{delta_t}{minimum time increment per-jump, defaults to sqrt(eps())}
}
\value{
A data frame with simulated data
}
\description{
Simulate n data sets or collections of sample paths from a multistate model.
If data = TRUE (the default) discretely observed sample paths are returned, possibly subject to measurement error.
If paths = FALSE (the default), continuous-time sample paths are not returned.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
d <- simulate(model = model_fit)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
