% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bic}
\alias{bic}
\title{BIC}
\usage{
bic(model_fitted, estimate_likelihood = TRUE, min_ess = 100, loglik = NULL)
}
\arguments{
\item{model_fitted}{A fitted multistate models object}

\item{estimate_likelihood}{logical; whether to estimate the log-likelihood, defaults to TRUE.}

\item{min_ess}{minimum effective sample size per subject, defaults to 100.}

\item{loglik}{value of the log-likelihood to use, if provided.}
}
\value{
An integer with calculated BIC
}
\description{
Bayesian Information Criterion, defined as -2\emph{log(L) + k}log(n),
where L is the likelihood and k is the number of consumed degrees of freedom.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
model_fitted <- fit(model = model_fit, verbose=TRUE, compute_vcov = TRUE, ess_target_initial = 50,
ascent_threshold = 0.2, stopping_threshold = 0.2, tol = 0.001)
bic(model_fitted = model_fitted)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
