% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_data}
\alias{collapse_data}
\title{Collapse subjects and recompute sampling weights}
\usage{
collapse_data(
  dat,
  multistatemodels,
  SamplingWeights = rep(1, length(unique(dat$id)))
)
}
\arguments{
\item{dat}{A data frame}

\item{multistatemodels}{Loaded multistatemodels Julia environment}

\item{SamplingWeights}{Sampling weights, defaults to 1 for each id in the data}
}
\value{
A 2-element list. The first element is a data frame and the second
is a vector of sampling weights
}
\description{
Collapse subjects to create an internal representation of a
dataset and optionally recompute a vector of sampling weights.
}
\examples{
multistatemodels <- multistatemodels_load()
df <- read.csv("~/derived_states_regen.csv")
d <- collapse_data(df, multistatemodels)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
