% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_data}
\alias{collapse_data}
\title{Collapse subjects and recompute sampling weights}
\usage{
collapse_data(dat, SamplingWeights = rep(1, length(unique(dat$id))))
}
\arguments{
\item{dat}{A data frame}

\item{SamplingWeights}{Sampling weights, defaults to 1 for each id in the data}
}
\value{
A 2-element list. The first element is a data frame and the second
is a vector of sampling weights
}
\description{
Collapse subjects to create an internal representation of a
dataset and optionally recompute a vector of sampling weights.
}
\examples{
d_collpased <- collapse_data(dat = illness_death_dat)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
