% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_hazard}
\alias{compute_hazard}
\title{Compute hazard}
\usage{
compute_hazard(t, model, hazard, multistatemodels, subj = 1)
}
\arguments{
\item{t}{time or vector of times}

\item{model}{multistate model}

\item{hazard}{string specifying the hazard, e.g., "h12" for the hazard for transitioning from state 1 to state 2.}

\item{multistatemodels}{Loaded multistatemodels Julia environment}

\item{subj}{subject id}
}
\value{
An integer or vector of integers with the hazard at the specified time(s)
}
\description{
Compute the hazard at time t
}
\examples{
compute_hazard(t=2, model=model, hazard="h12", multistatemodels=multistatemodels)
}
