% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_hazard}
\alias{compute_hazard}
\title{Compute hazard}
\usage{
compute_hazard(t, model, hazard, subj = 1)
}
\arguments{
\item{t}{time or vector of times}

\item{model}{multistate model}

\item{hazard}{string specifying the hazard, e.g., "h12" for the hazard for transitioning from state 1 to state 2.}

\item{subj}{subject id}
}
\value{
An integer or vector of integers with the hazard at the specified time(s)
}
\description{
Compute the hazard at time t
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
compute_hazard(t=0.5, model=model_fit, hazard = "h12")
\dontshow{
JuliaConnectoR::stopJulia()
}
}
