% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cumulative_incidence}
\alias{cumulative_incidence}
\title{Cumulative incidence}
\usage{
cumulative_incidence(t, model, subj = 1)
}
\arguments{
\item{t}{time or a vector of times}

\item{model}{multistate model}

\item{subj}{subject id}
}
\value{
A matrix with the cumulative incidence where columns represent each
possible transition and rows represent t
}
\description{
Compute the cumulative incidence for each possible transition as a function of time since state entry.
Assumes the subject starts their observation period at risk and saves cumulative incidence at the supplied vector of times, t.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
cumulative_incidence(t=0.5, model=model_fit)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
