% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_paths}
\alias{draw_paths}
\title{Draw paths}
\usage{
draw_paths(model, min_ess = 100, paretosmooth = TRUE, return_logliks = FALSE)
}
\arguments{
\item{model}{multistate model}

\item{min_ess}{minimum effective sample size, defaults to 100.}

\item{paretosmooth}{logical; pareto smooth importance weights, defaults to TRUE unless min_ess < 25.}

\item{return_logliks}{logical; defaults to FALSE}
}
\value{
A Julia environment with a named Tuple of sample paths
}
\description{
Draw sample paths conditional on the data.
Require that the minimum effective sample size is greater than min_ess.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
draw_paths(model = model_fit)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
