% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_loglik}
\alias{estimate_loglik}
\title{Estimate log-likelihood}
\usage{
estimate_loglik(model, min_ess = 100)
}
\arguments{
\item{model}{multistate model}

\item{min_ess}{minimum effective sample size, defaults to 100.}
}
\value{
A Julia environment with a named Tuple
}
\description{
Estimate the log marginal likelihood for a fitted multistate model.
Require that the minimum effective sample size per subject is greater than min_ess.
}
