% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_loglik}
\alias{estimate_loglik}
\title{Estimate log-likelihood}
\usage{
estimate_loglik(model, min_ess = 100)
}
\arguments{
\item{model}{multistate model}

\item{min_ess}{minimum effective sample size, defaults to 100.}
}
\value{
A Julia environment with a named Tuple
}
\description{
Estimate the log marginal likelihood for a fitted multistate model.
Require that the minimum effective sample size per subject is greater than min_ess.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
estimate_loglik(model = model_fit)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
