% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_surrogate}
\alias{fit_surrogate}
\title{Fit a Markov surrogate model}
\usage{
fit_surrogate(
  model,
  surrogate_parameters = NULL,
  surrogate_constraints = NULL,
  crude_inits = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{model}{A model created in the multistatemodel function}

\item{surrogate_parameters}{Surrogate parameters}

\item{surrogate_constraints}{Surrogate constraints on model parameters}

\item{crude_inits}{logical; Defaults to TRUE}

\item{verbose}{logical; print messages, defaults to TRUE}
}
\value{
A Julia environment with a fitted Markov surrogate model
}
\description{
Fit a Markov surrogate model
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
model_fitted_surr <- fit_surrogate(model = model_fit)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
