% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ConvergenceRecords}
\alias{get_ConvergenceRecords}
\title{Get Convergence Records}
\usage{
get_ConvergenceRecords(model_fitted, multistatemodels)
}
\arguments{
\item{model_fitted}{A fitted multistate models object}

\item{multistatemodels}{Loaded multistatemodels Julia environment}
}
\value{
A Julia environment with status, candidate solution, algorithm,
convergence records, and work counters
}
\description{
Return the convergence records for the fit.
}
\examples{
get_ConvergenceRecords(model_fitted=model_fitted, multistatemodels=multistatemodels)
}
