% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_loglik}
\alias{get_loglik}
\title{Get Log-Likelihood}
\usage{
get_loglik(model_fitted, ll = "loglik")
}
\arguments{
\item{model_fitted}{A fitted multistate models object}

\item{ll}{string; one of "loglik" (default) for the observed data log-likelihood or
"subj_lml" for log marginal likelihood at the subject level}
}
\value{
An integer with the log-likelihood at the MLE
}
\description{
Return the log-likelihood at the maximum likelihood estimates.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
model_fitted <- fit(model = model_fit, verbose=TRUE, compute_vcov = TRUE, ess_target_initial = 50,
ascent_threshold = 0.2, stopping_threshold = 0.2, tol = 0.001)
get_loglik(model_fitted=model_fitted)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
