% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_parameters}
\alias{get_parameters}
\title{Get Parameters}
\usage{
get_parameters(model_fitted, multistatemodels)
}
\arguments{
\item{model_fitted}{A fitted multistate models object}

\item{multistatemodels}{Loaded multistatemodels Julia environment}
}
\value{
A matrix of model parameters
}
\description{
Return the model parameters
}
\examples{
get_parameters(model_fitted=model_fitted, multistatemodels=multistatemodels)
}
