% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Hazard}
\alias{Hazard}
\title{Specify hazard}
\usage{
Hazard(
  formula,
  statefrom,
  stateto,
  family = c("exp", "wei", "sp"),
  degree = 3,
  natural_spline = TRUE,
  extrapolation = "linear",
  monotone = 0,
  knots = NULL,
  boundaryknots = NULL
)
}
\arguments{
\item{formula}{Formula for log-hazard. Covariates have a multiplicative
effect on the baseline cause specific hazard.
Must be specified with "0 ~" on the left hand side.}

\item{statefrom}{Integer specifying the origin state}

\item{stateto}{Integer specifying the destination state}

\item{family}{string; One of "exp" or "wei" for exponential or Weibull cause-specific
baseline hazard functions, or "sp" for a semi-parametric spline basis for the
baseline hazard (defaults to M-splines).}

\item{degree}{For family="sp": degree of the spline polynomial basis}

\item{natural_spline}{For family="sp": logical; Restrict the second derivative to zero at the boundaries, defaults to TRUE.}

\item{extrapolation}{For family="sp": string; Either "linear" or "flat", defaults to "linear"}

\item{monotone}{For family="sp": 0, -1, or 1 for non-monotone, monotone decreasing, or monotone increasing.}

\item{knots}{For family="sp": Optional vector of knots. Defaults to the range of sojourns in
the data with no interior knots if not supplied.}

\item{boundaryknots}{For family="sp": Optional vector of boundary knots,
defaults to the range of possible sojourn times if not supplied.}
}
\value{
An environment with a Julia hazard
}
\description{
Specify a parametric or semi-parametric baseline cause-specific
hazard function.
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
\dontshow{
JuliaConnectoR::stopJulia()
}
}
