% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_constraints}
\alias{make_constraints}
\title{Make parameter constraints}
\usage{
make_constraints(cons, lcons, ucons)
}
\arguments{
\item{cons}{string or vector of strings, name of constraints}

\item{lcons}{lcons}

\item{ucons}{ucons}
}
\value{
A Julia environment with a named Tuple of constraints
}
\description{
Make parameter constraints
}
\examples{
constraints_weibull <- make_constraints(cons = "h12_shape", lcons = 0, ucons = 0)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
