% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_summary}
\alias{model_summary}
\title{Summary of model output}
\usage{
model_summary(
  model_fitted,
<<<<<<< HEAD
=======
  multistatemodels,
>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
  confidence_level = 0.95,
  estimate_likelihood = FALSE,
  min_ess = 100
)
}
\arguments{
\item{model_fitted}{fitted multistate models object}

<<<<<<< HEAD
=======
\item{multistatemodels}{Loaded multistatemodels Julia environment}

>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
\item{confidence_level}{confidence level of the confidence intervals, defaults to 0.95}

\item{estimate_likelihood}{logical; estimate likelihood, defaults to FALSE}

\item{min_ess}{minimum effective sample size, defaults to 100.}
}
\value{
A Julia environment with a named Tuple
Estimate, SE, upper and lower confidence limits for each hazard parameter
loglik, AIC, BIC, MCSE_loglik
}
\description{
Summary of model output
}
\examples{
<<<<<<< HEAD
model_summary(model_fitted=model_fitted, estimate_likelihood=TRUE)
=======
model_summary(model_fitted=model_fitted, multistatemodels=multistatemodels, estimate_likelihood=TRUE)
>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
}
