% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_summary}
\alias{model_summary}
\title{Summary of model output}
\usage{
model_summary(
  model_fitted,
  confidence_level = 0.95,
  estimate_likelihood = FALSE,
  min_ess = 100
)
}
\arguments{
\item{model_fitted}{fitted multistate models object}

\item{confidence_level}{confidence level of the confidence intervals, defaults to 0.95}

\item{estimate_likelihood}{logical; estimate likelihood, defaults to FALSE}

\item{min_ess}{minimum effective sample size, defaults to 100.}
}
\value{
A Julia environment with a named Tuple
Estimate, SE, upper and lower confidence limits for each hazard parameter
loglik, AIC, BIC, MCSE_loglik
}
\description{
Summary of model output
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
model_fitted <- fit(model = model_fit, verbose=TRUE, compute_vcov = TRUE, ess_target_initial = 50,
ascent_threshold = 0.2, stopping_threshold = 0.2, tol = 0.001)
model_summary(model_fitted = model_fitted, estimate_likelihood = TRUE)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
