% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistatemodel}
\alias{multistatemodel}
\title{Construct a multistate model}
\usage{
multistatemodel(
  hazard,
  data,
  SamplingWeights = NULL,
  CensoringPatterns = NULL,
  verbose = FALSE
)
}
\arguments{
\item{hazard}{A vector of cause specific hazards}

\item{data}{Data frame}

\item{SamplingWeights}{Sampling weights}

\item{CensoringPatterns}{Censoring patterns}

\item{verbose}{logical; print messages, defaults to FALSE}
}
\value{
A Julia MultistateModel environment that can be used for simulation and inference
}
\description{
Construct a multistate model
}
\examples{
h12 <- Hazard(formula = 0~1, statefrom = 1, stateto=2, family="wei")
h13 <- Hazard(formula = 0~1, statefrom = 1, stateto=3, family="wei")
h23 <- Hazard(formula = 0~1, statefrom = 2, stateto=3, family="wei")
model_fit <- multistatemodel(hazard = c(h12, h13, h23), data=illness_death_dat)
\dontshow{
JuliaConnectoR::stopJulia()
}
}
