% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistatemodels_setup}
\alias{multistatemodels_setup}
\title{Check Julia setup}
\usage{
multistatemodels_setup(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
Check Julia setup
}
\examples{
multistatemodels_setup()
}
