% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multistatemodels_setup}
\alias{multistatemodels_setup}
\title{Setup Multistate Models with Julia packages}
\usage{
multistatemodels_setup()
}
\description{
This function should be run any time there are updates to the Julia  packages
used (specifically MultistateModels).
This allows the necessary Julia packages to be updated to use in R.
}
\examples{
multistatemodels_setup()
}
