% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_parameters}
\alias{set_parameters}
\title{Set parameters}
\usage{
<<<<<<< HEAD
set_parameters(model, newvalues)
=======
set_parameters(model, multistatemodels, newvalues)
>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
}
\arguments{
\item{model}{multistate model}

<<<<<<< HEAD
=======
\item{multistatemodels}{Loaded multistatemodels Julia environment}

>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
\item{newvalues}{A list, each element of the list is new parameters for each hazard}
}
\description{
Set model parameters given a vector of values.
Copies newvalues to model.parameters.
}
\examples{
<<<<<<< HEAD
set_parameters(model, newvalues = list(h12 = c(log(1.5), log(1)),
=======
set_parameters(model, multistatemodels, newvalues = list(h12 = c(log(1.5), log(1)),
>>>>>>> 58defaf854b74cbdc24e43f36f206de956a16d51
h23=c(log(2/3), log(1)), h24=c(log(1), log(1.25)), h45=c(log(1), log(1.25))))
}
